requirements.txt
celery
django
django-elasticsearch-dsl
djangorestframework
pillow
psycopg2
redis